+++ core/admin/gis_file_create.py
@@ -100,7 +100,7 @@ class GwGisFileCreate:
                                 level = old_level
                     rectangle = tools_gw._get_extent_parameters(schema)
                     # Add project layer
-                    tools_gw.add_layer_database(layer['tableName'], layer['geomField'], layer['tableId'], levels[0], levels[1] if len(levels) > 1 else None, style_id='-1', alias=layer['layerName'],
+                    tools_gw.add_layer_database(layer['tableName'], layer['geomField'], layer['tableId'], levels[0], levels[1] if len(levels) > 1 else None, alias=layer['layerName'],
                                                  sub_sub_group=levels[2] if len(levels) > 2 else None, schema=layer['tableSchema'], visibility=template.get('visibility'), auth_id=auth_id,
                                                  extent=rectangle, passwd=self.layer_source['password'] if export_passwd else None, create_project=True, force_create_group=False,
                                                  properties=properties)
@@ -108,7 +108,7 @@ class GwGisFileCreate:
         # Hide hidden group
         tools_gw.hide_group_from_toc('HIDDEN')
         root.findGroup('HIDDEN').setItemVisibilityChecked(False)
-        
+
         # Set project CRS
         project.setCrs(QgsCoordinateReferenceSystem(auth_id))
 
@@ -120,7 +120,7 @@ class GwGisFileCreate:
         for layer in layers:
             if isinstance(layer, QgsLayerTreeLayer):
                 layer.setExpanded(False)
-        
+
         # Collapse all groups
         groups = root.findGroups()
         for group in groups:
@@ -144,7 +144,7 @@ class GwGisFileCreate:
             tools_qgis.show_warning(msg, parameter=qgs_path)
             return False, qgs_path
 
-    # region private functions 
+    # region private functions
 
     def _get_database_parameters(self, schema):
         """ Get database parameters from layer source """