+++ core/utils/tools_gw.py
@@ -25,6 +25,7 @@ if 'nt' in sys.builtin_module_names:
 from collections import OrderedDict
 from functools import partial
 from datetime import datetime
+from urllib.parse import urlencode
 
 from qgis.PyQt.QtCore import Qt, QStringListModel, QVariant, QDate, QSettings, QLocale, QRegularExpression, \
     QItemSelectionModel, QTimer
@@ -39,7 +40,8 @@ from qgis.core import Qgis, QgsProject, QgsPointXY, QgsVectorLayer, QgsField, Qg
     QgsFeatureRequest, QgsSimpleFillSymbolLayer, QgsRendererCategory, QgsCategorizedSymbolRenderer, \
     QgsCoordinateTransform, QgsCoordinateReferenceSystem, QgsVectorFileWriter, QgsCoordinateTransformContext, \
     QgsFieldConstraints, QgsEditorWidgetSetup, QgsRasterLayer, QgsGeometry, QgsExpression, QgsRectangle, \
-    QgsEditFormConfig, QgsSymbolLayer, QgsProperty, QgsSimpleLineSymbolLayer, QgsSimpleMarkerSymbolLayer, QgsStyle
+    QgsEditFormConfig, QgsSymbolLayer, QgsProperty, QgsSimpleLineSymbolLayer, QgsSimpleMarkerSymbolLayer, QgsStyle, \
+    QgsDataSourceUri
 from qgis.gui import QgsDateTimeEdit, QgsRubberBand, QgsExpressionSelectionDialog
 
 from ..models.cat_feature import GwCatFeature
@@ -864,7 +866,7 @@ def set_completer_feature_id(widget, feature_type, viewname):
         completer.setModel(model)
 
 
-def add_layer_database(tablename=None, the_geom="the_geom", field_id="id", group="GW Layers", sub_group=None, style_id="-1", alias=None, sub_sub_group=None, schema=None,
+def add_layer_database(tablename=None, the_geom="the_geom", field_id="id", group="GW Layers", sub_group=None, alias=None, sub_sub_group=None, schema=None,
                         visibility=None, auth_id=None, extent=None, passwd=None, create_project=False, force_create_group=False, properties=None):
     """
     Put selected layer into TOC
@@ -873,7 +875,6 @@ def add_layer_database(tablename=None, the_geom="the_geom", field_id="id", group
         :param field_id: Field id of the table (String)
         :param child_layers: List of layers (StringList)
         :param group: Name of the group that will be created in the toc (String)
-        :param style_id: Id of the style we want to load (integer or String)
         :param alias: Alias of the layer (String)
         :param sub_sub_group: Sub-sub-group of the layer (String)
         :param schema: Schema of the layer (String)
@@ -917,16 +918,16 @@ def add_layer_database(tablename=None, the_geom="the_geom", field_id="id", group
         if alias:
             tablename = alias
         layer = QgsRasterLayer(connString, tablename)
-        tools_qgis.add_layer_to_toc(layer, group, sub_group, create_groups=create_groups)
+        tools_qgis.add_layer_to_toc(layer, group, sub_group, create_groups=create_groups, custom_properties={"gw_id": tablename_og})
 
     else:
         if alias:
             tablename = alias
         layer = QgsVectorLayer(uri.uri(), f'{tablename}', 'postgres')
-        tools_qgis.add_layer_to_toc(layer, group, sub_group, create_groups=create_groups, sub_sub_group=sub_sub_group)
+        tools_qgis.add_layer_to_toc(layer, group, sub_group, create_groups=create_groups, sub_sub_group=sub_sub_group, custom_properties={"gw_id": tablename_og})
 
         # Apply styles to layer
-        if style_id in (None, "-1") and schema_name != 'cm':
+        if schema_name != 'cm':
             set_layer_styles(tablename_og, layer, schema_name if schema_name != "am" else None)
             # Get addparam from sys_table
             sql = f"SELECT addparam FROM sys_table WHERE id = '{tablename_og}'"
@@ -973,6 +974,248 @@ def add_layer_database(tablename=None, the_geom="the_geom", field_id="id", group
 
     if create_project is False:
         global_vars.iface.mapCanvas().refresh()
+
+
+def add_layer_provider(gw_id: str, cfg, group="GW Layers", sub_group=None, alias=None, sub_sub_group=None,
+                        schema=None, visibility=None, force_create_group=False):
+    """
+    Put selected layer into TOC from provider configuration
+        :param gw_id: Giswater ID (String)
+        :param cfg: Provider configuration (dict)
+        :param group: Name of the group that will be created in the toc (String)
+        :param sub_group: Sub-group of the layer (String)
+        :param alias: Alias of the layer (String)
+        :param sub_sub_group: Sub-sub-group of the layer (String)
+        :param schema: Schema of the layer (String)
+        :param visibility: Visibility of the layer (Boolean)
+        :param force_create_group: Force create group (Boolean)
+    """
+    schema_name = tools_db.dao_db_credentials['schema'].replace('"', '') if schema is None else schema
+    gw_id_og = gw_id
+
+    provider = cfg["provider"]
+    layer_type = cfg["layer_type"]
+
+    uri = build_uri(gw_id, provider, cfg)
+    if uri is None:
+        return
+
+    if alias:
+        gw_id = alias
+
+    if layer_type == "raster":
+        layer = QgsRasterLayer(uri, gw_id, provider)
+
+    elif layer_type == "vector":
+        layer = QgsVectorLayer(uri, gw_id, provider)
+
+    else:
+        raise ValueError(f"Unsupported layer_type: {layer_type}")
+
+    if not layer.isValid():
+        raise Exception(
+            f"Invalid layer:\n"
+            f"provider={provider}\n"
+            f"uri={uri}"
+        )
+
+    if force_create_group:
+        create_groups = True
+    else:
+        create_groups = get_config_parser("system", "force_create_qgis_group_layer", "user", "init", prefix=False)
+        create_groups = tools_os.set_boolean(create_groups, default=False)
+
+    if sub_group:
+        sub_group = sub_group.capitalize()
+    if sub_sub_group:
+        sub_sub_group = sub_sub_group.capitalize()
+
+    tools_qgis.add_layer_to_toc(layer, group, sub_group, create_groups=create_groups, sub_sub_group=sub_sub_group, custom_properties={"gw_id": gw_id_og})
+
+    # Apply styles to layer
+    set_layer_styles(gw_id_og, layer, schema_name)
+    # Get addparam from sys_table
+    sql = f"SELECT addparam FROM sys_table WHERE id = '{gw_id_og}'"
+    row = tools_db.get_row(sql)
+    if row:
+        addparam = row[0]
+        if addparam and addparam.get('refreshSymbology'):
+            renderer = layer.renderer()
+            if isinstance(renderer, QgsCategorizedSymbolRenderer):
+                refresh_categorized_layer_symbology_classes(layer, addparam)
+
+    if visibility is not None:
+        if visibility is False:
+            tools_qgis.set_layer_visible(layer, recursive=False, visible=False)
+
+
+def build_uri(gw_id: str, provider: str, cfg: dict) -> Optional[str]:
+
+    # direct override
+    raw_uri = cfg.get("raw_uri")
+    if raw_uri:
+        return raw_uri
+
+    builders = {
+        "wms": build_generic_network_uri,
+        "wfs": build_generic_network_uri,
+        "oapif": partial(build_generic_network_uri, qgis_style=True),
+        "xyz": build_generic_network_uri,
+        "geojson": partial(build_geojson_uri, gw_id=gw_id)
+    }
+
+    builder = builders.get(provider)
+
+    if not builder:
+        tools_qgis.show_warning(f"No URI builder for provider: {provider}")
+        return None
+
+    return builder(cfg)
+
+
+def build_geojson_uri(gw_id: str, cfg: dict) -> Optional[str]:
+    """
+    Build a GeoJSON URI from a provider configuration
+        :param gw_id: Giswater ID (String)
+        :param cfg: Provider configuration (dict)
+        :return: URI (String)
+
+    Example of supported format:
+
+    {
+        "provider": "geojson",
+        "source": {
+            "path": "/tmp/test.geojson"
+        }
+    }
+
+    {
+        "provider": "geojson",
+        "source": {
+            "url": "https://server/data.geojson"
+        }
+    }
+
+    {
+        "provider": "geojson",
+        "source": {
+            "geojson": {...}
+        }
+    }
+    """
+
+    source = cfg.get("source", {})
+
+    # local file
+    if "path" in source:
+        return source["path"]
+
+    # remote URL
+    if "url" in source:
+        return source["url"]
+
+    # inline geojson object
+    if "geojson" in source:
+
+        geojson_data = source["geojson"]
+
+        geojson_str = json.dumps(geojson_data)
+
+        timestamp = datetime.now().strftime("%Y%m%d%H%M%S")
+        file_name = f"{gw_id}_{timestamp}"
+
+        vsipath = f"/vsimem/{file_name}.geojson"
+
+        gdal.FileFromMemBuffer(
+            vsipath,
+            geojson_str.encode("utf-8")
+        )
+
+        return vsipath
+
+    return None
+
+
+def build_generic_network_uri(cfg: dict, qgis_style: bool = False) -> str:
+    """
+    Build a network URI from a provider configuration
+        :param cfg: Provider configuration (dict)
+        :param qgis_style: If True, return the URI in QGIS style (Boolean)
+        :return: URI (String)
+        
+    Examples of supported formats:
+
+    WMS:
+    {
+        "source": {
+            "url": "https://example.com/wms"
+        },
+        "provider": "wms",
+        "layer_type": "raster",
+        "uri_params": {
+            "format": "image/png",
+            "layers": "example_layer",
+            "styles": "",
+            "dpiMode": 7,
+            "featureCount": 10,
+            "contextualWMSLegend": "0"
+        }
+    }
+
+    OAPIF (OGC API Features):
+    {
+        "source": {
+            "url": "https://example.com/oapif"
+        },
+        "provider": "oapif",
+        "layer_type": "vector",
+        "uri_params": {
+            "srsname": "OGC:CRS84",
+            "version": "auto",
+            "typename": "example_collection",
+            "pagingEnabled": "default",
+            "restrictToRequestBBOX": "1",
+            "preferCoordinatesForWfsT11": "false"
+        }
+    }
+    """
+
+    source = cfg.get("source", {})
+    uri_params = cfg.get("uri_params", {}).copy()
+
+    network = cfg.get("network", {})
+
+    # merge source
+    uri_params.update(source)
+
+    # custom headers
+    headers = network.get("headers", {})
+    if headers:
+
+        header_string = "&".join(
+            f"{k}:{v}"
+            for k, v in headers.items()
+        )
+
+        uri_params["http-header"] = header_string
+
+    uri = QgsDataSourceUri()
+
+    for key, value in uri_params.items():
+        if value is None:
+            continue
+
+        uri.setParam(str(key), str(value))
+
+    if qgis_style:
+        return uri.uri(False)
+
+    encoded_uri = uri.encodedUri()
+
+    if isinstance(encoded_uri, (bytes, bytearray)):
+        return encoded_uri.decode()
+
+    return str(encoded_uri)
 
 
 def refresh_categorized_layer_symbology_classes(layer, addparam=None):
@@ -1144,6 +1387,12 @@ def validate_qml(qml_content):
 
 
 def set_layer_styles(tablename, layer, schema_name):
+    # Save current layer custom properties and remove them to avoid QML overwrite them
+    custom_properties = {}
+    layer_custom_properties = layer.customPropertyKeys()
+    for cp_key in layer_custom_properties:
+        custom_properties[cp_key] = layer.customProperty(cp_key)
+
     body = f'$${{"data":{{"layername":"{tablename}"}}}}$$'
     json_return = execute_procedure('gw_fct_getstyle', body, schema_name=schema_name)
     if json_return is None or json_return['status'] == 'Failed':
@@ -1166,6 +1415,10 @@ def set_layer_styles(tablename, layer, schema_name):
                 # set new style as current
                 style_manager.setCurrentStyle(style_name)
                 tools_qgis.create_qml(layer, qml)
+
+    # Restore actual custom properties
+    for cp_key, cp_value in custom_properties.items():
+        layer.setCustomProperty(cp_key, cp_value)
 
 
 def add_layer_temp(dialog, data, layer_name, force_tab=True, reset_text=True, tab_idx=1, del_old_layers=True,
@@ -1206,11 +1459,13 @@ def add_layer_temp(dialog, data, layer_name, force_tab=True, reset_text=True, ta
                 aux_layer_name = str(k)
 
             if del_old_layers:
-                tools_qgis.remove_layer_from_toc(aux_layer_name, group)
+                tools_qgis.remove_layer(custom_properties={"gw_id": aux_layer_name}, group_name=group)
 
             # Use GeoJSON approach like manage_json_return
             geojson_str = json.dumps(data[k])
-            vsipath = f"/vsimem/{aux_layer_name}.geojson"
+            timestamp = datetime.now().strftime("%Y%m%d%H%M%S")
+            file_name = f"{aux_layer_name}_{timestamp}"
+            vsipath = f"/vsimem/{file_name}.geojson"
             gdal.FileFromMemBuffer(vsipath, geojson_str)
             v_layer = QgsVectorLayer(vsipath, aux_layer_name, 'ogr')
 
@@ -1223,12 +1478,7 @@ def add_layer_temp(dialog, data, layer_name, force_tab=True, reset_text=True, ta
             geometry_type = v_layer.geometryType()
 
             # Add layer to project and group
-            QgsProject.instance().addMapLayer(v_layer, False)
-            root = QgsProject.instance().layerTreeRoot()
-            my_group = root.findGroup(group)
-            if my_group is None:
-                my_group = root.insertGroup(0, group)
-            my_group.insertLayer(i, v_layer)
+            tools_qgis.add_layer_to_toc(v_layer, group=group, create_groups=True, custom_properties={"gw_id": aux_layer_name})
 
             # Increase iterator
             i = i + 1
@@ -1580,7 +1830,7 @@ def load_layer_in_hidden_group(layer_name, key_column, add_to_toc=True):
         layer = QgsVectorLayer(uri.uri(False), layer_name, "postgres")
         if layer.isValid():
             if add_to_toc:
-                tools_qgis.add_layer_to_toc(layer, group="HIDDEN", create_groups=True)
+                tools_qgis.add_layer_to_toc(layer, group="HIDDEN", create_groups=True, custom_properties={"gw_id": table})
             return layer
 
     # Select the layer called 've_node'
@@ -1706,12 +1956,7 @@ def fill_layer_temp(virtual_layer, data, layer_type, counter, group='GW Temporal
 
     # Commit changes
     virtual_layer.commitChanges()
-    QgsProject.instance().addMapLayer(virtual_layer, False)
-    root = QgsProject.instance().layerTreeRoot()
-    my_group = root.findGroup(group)
-    if my_group is None:
-        my_group = root.insertGroup(0, group)
-    my_group.insertLayer(sort_val, virtual_layer)
+    tools_qgis.add_layer_to_toc(virtual_layer, group=group, create_groups=True, custom_properties={"gw_id": virtual_layer.name()})
 
 
 def enable_widgets(dialog, result, enable):
@@ -3860,7 +4105,7 @@ def manage_json_return(json_result, sql, rubber_band=None, i=None):  # noqa: C90
                 layer_name = f'{key}'
                 if value.get('layerName'):
                     layer_name = value['layerName']
-                tools_qgis.remove_layer_from_toc(layer_name, 'GW Temporal Layers')
+                tools_qgis.remove_layer(custom_properties={"gw_id": layer_name}, group_name='GW Temporal Layers')
 
                 if 'features' not in value:
                     continue
@@ -3869,19 +4114,16 @@ def manage_json_return(json_result, sql, rubber_band=None, i=None):  # noqa: C90
 
                 # Get values for create and populate layer
                 geojson_str = json.dumps(value)
-                vsipath = f"/vsimem/{layer_name}.geojson"
+                timestamp = datetime.now().strftime("%Y%m%d%H%M%S")
+                file_name = f"{layer_name}_{timestamp}"
+                vsipath = f"/vsimem/{file_name}.geojson"
                 gdal.FileFromMemBuffer(vsipath, geojson_str)
                 v_layer = QgsVectorLayer(vsipath, layer_name, 'ogr')
                 if not v_layer.crs().isValid():
                     v_layer.setCrs(QgsCoordinateReferenceSystem('EPSG:4326'))
                 v_layer.updateExtents()
                 geometry_type = v_layer.geometryType()
-                QgsProject.instance().addMapLayer(v_layer, False)
-                root = QgsProject.instance().layerTreeRoot()
-                my_group = root.findGroup('GW Temporal Layers')
-                if my_group is None:
-                    my_group = root.insertGroup(0, 'GW Temporal Layers')
-                my_group.insertLayer(i, v_layer)
+                tools_qgis.add_layer_to_toc(v_layer, group='GW Temporal Layers', create_groups=True, custom_properties={"gw_id": layer_name})
 
                 # Increase iterator
                 i = i + 1
@@ -3924,6 +4166,12 @@ def manage_json_return(json_result, sql, rubber_band=None, i=None):  # noqa: C90
                     v_layer.renderer().symbol().setOpacity(opacity)
 
                 elif return_manager_style['style'] == 'qml':
+                    # Save current layer custom properties and remove them to avoid QML overwrite them
+                    custom_properties = {}
+                    layer_custom_properties = v_layer.customPropertyKeys()
+                    for cp_key in layer_custom_properties:
+                        custom_properties[cp_key] = v_layer.customProperty(cp_key)
+
                     style_id = return_manager_style['id']
                     extras = f'"style_id":"{style_id}", "layername":"{layer_name}"'
                     body = create_body(extras=extras)
@@ -3952,6 +4200,10 @@ def manage_json_return(json_result, sql, rubber_band=None, i=None):  # noqa: C90
                                 # set new style as current
                                 style_manager.setCurrentStyle(style_name)
                                 tools_qgis.create_qml(v_layer, qml)
+
+                    # Restore actual custom properties
+                    for cp_key, cp_value in custom_properties.items():
+                        v_layer.setCustomProperty(cp_key, cp_value)
 
                 elif return_manager_style['style'] == 'unique':
                     color = return_manager_style['values']['color']
@@ -4331,8 +4583,7 @@ def manage_layer_manager(json_result, sql=None):
                         group = lyr[layer_name]['group_layer']
                     else:
                         group = "GW Layers"
-                    style_id = lyr[layer_name]['style_id']
-                    add_layer_database(layer_name, the_geom, field_id, group=group, style_id=style_id, alias=lyr.get(layer_name).get('alias'))
+                    add_layer_database(layer_name, the_geom, field_id, group=group, alias=lyr.get(layer_name).get('alias'))
                 tools_qgis.set_layer_visible(layer)
 
         # force reload dataProvider in order to reindex.